refactor: relocate TextFieldWithLabel to separate files
